requirements.txt
streamlit>=1.36.0
pandas>=2.2.2
numpy>=1.26.4
openpyxl>=3.1.2
python-dateutil>=2.9.0.post0
Unidecode>=1.3.8
python-calamine>=0.2.0
